fix: workflow
