+++ .github/scripts/documentation.js
@@ -1,30 +1,27 @@
 /**
- * Documentation Update Utilities
+ * Chart Documentation Utilities
  * 
- * This module provides functions for automating chart documentation updates:
- * - Installing helm-docs
- * - Generating documentation with helm-docs
- * - Committing changes to pull requests
+ * This module provides functions for automating Helm chart documentation updates,
+ * including installing the helm-docs tool and generating updated documentation.
+ * It handles the process of generating and committing documentation changes to
+ * pull requests, ensuring consistency across chart documentation.
  * 
  * @module documentation
+ * @author AXIVO
+ * @license BSD-3-Clause
  */
 
 const os = require('os');
-const gitSignedCommit = require('./git-signed-commit');
+const commit = require('./git-signed-commit');
 const utils = require('./utils');
 
 /**
- * Configuration constants for Documentation module
- * Contains settings for helm-docs install and other Git-related parameters
- */
-const CONFIG = {
-  helmDocs: {
-    baseUrl: 'https://github.com/norwoodj/helm-docs/releases/download'
-  }
-};
-
-/**
  * Installs the helm-docs package for generating Helm chart documentation
+ * 
+ * This function downloads and installs the helm-docs package from the official
+ * GitHub repository. It handles the process of getting the package, installing it
+ * using system package management, and making it available for chart documentation
+ * generation.
  * 
  * @param {Object} params - Function parameters
  * @param {Object} params.core - GitHub Actions Core API for logging and output
@@ -41,7 +38,8 @@ async function installHelmDocs({
     const tmpDir = os.tmpdir();
     const packageFile = `helm-docs_${version}_Linux_x86_64.deb`;
     const packagePath = [tmpDir, packageFile].join('/');
-    const packageUrl = [CONFIG.helmDocs.baseUrl, `v${version}`, packageFile].join('/');
+    const packageBaseUrl = 'https://github.com/norwoodj/helm-docs/releases/download';
+    const packageUrl = [packageBaseUrl, `v${version}`, packageFile].join('/');
     core.info(`Installing helm-docs v${version}...`);
     const runSudo = async (args) => (await exec.getExecOutput('sudo', args)).stdout.trim();
     await runSudo(['wget', '-qP', tmpDir, '-t', '10', '-T', '60', packageUrl]);
@@ -54,6 +52,10 @@ async function installHelmDocs({
 
 /**
  * Updates documentation in a pull request by generating docs and committing changes
+ * 
+ * This function runs the helm-docs tool to automatically update all chart documentation
+ * based on the Chart.yaml files and templates. It detects changes to documentation files
+ * and creates a signed commit with these changes to the pull request branch.
  * 
  * @param {Object} options - Options for updating documentation
  * @param {Object} options.github - GitHub API client
@@ -84,10 +86,10 @@ async function updateDocumentation({
       core.info('No file changes detected, documentation is up to date');
       return;
     }
-    const { additions, deletions } = await gitSignedCommit.getGitStagedChanges(runGit);
+    const { additions, deletions } = await commit.getGitStagedChanges(runGit);
     if (additions.length + deletions.length > 0) {
       const currentHead = await runGit(['rev-parse', 'HEAD']);
-      await gitSignedCommit.createSignedCommit({
+      await commit.createSignedCommit({
         github,
         context,
         core,
@@ -106,6 +108,13 @@ async function updateDocumentation({
   }
 }
 
+/**
+ * Exports the module's functions
+ * 
+ * This module exports functions for installing helm-docs and updating chart
+ * documentation. These functions are designed to be used within GitHub Actions
+ * workflows to automate the process of keeping chart documentation up-to-date.
+ */
 module.exports = {
   installHelmDocs,
   updateDocumentation
