+++ .github/workflows/release.yml
@@ -58,14 +58,14 @@ jobs:
         uses: actions/github-script@v7
         with:
           script: |
-            const { processReleases } = require('./.github/scripts/chart.js');
+            const { processReleases } = require('./.github/scripts/release.js');
             await processReleases({ github, context, core, exec });
 
       - name: Generate charts index
         uses: actions/github-script@v7
         with:
           script: |
-            const { generateIndex } = require('./.github/scripts/chart.js');
+            const { generateIndex } = require('./.github/scripts/release.js');
             await generateIndex({ context, core });
 
       - name: Setup release build environment
